Moved Entry object to parent package
