+++ concourse-plugin-core/src/main/java/com/cinchapi/concourse/plugin/data/DatasetEntry.java
@@ -1,4 +1,4 @@
-package com.cinchapi.concourse.plugin.data.entry;
+package com.cinchapi.concourse.plugin.data;
 
 public class DatasetEntry<E, A, V> {
