+++ concourse-plugin-core/src/test/java/com/cinchapi/concourse/plugin/data/DatasetTest.java
@@ -23,7 +23,6 @@ import java.util.Set;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.cinchapi.concourse.plugin.data.entry.DatasetEntry;
 import com.cinchapi.concourse.test.ConcourseBaseTest;
 import com.cinchapi.concourse.util.Random;
